% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdqpt.r
\name{hdqpt}
\alias{hdqpt}
\title{High dimensional quantile partial test}
\usage{
hdqpt(
  x,
  u,
  y,
  tau = 0.5,
  coef = NULL,
  decorrelate = TRUE,
  W = NULL,
  method = c("hqreg", "rqPen"),
  seed = 1
)
}
\arguments{
\item{x}{The design matrix, which is a n by p matrix.}

\item{u}{The control factors, which is a n by q matrix.}

\item{y}{The outcome.}

\item{tau}{The quantile.}

\item{coef}{The coefficient of control factors, the first element corresponding to intercept. If not provided, we will first estimate according to the user's choice of estimation method, and then test process.}

\item{decorrelate}{An indicator of whether do decorrelate test.}

\item{W}{The decorrelate matrix, should be a q by p matrix when provided. Necessary when decorrelate is TRUE. If not provided, we will estimate it using cv.glmnet function from glmnet.}

\item{method}{Method used to estimator the coefficient of control factors when it is not provided. "hqreg" means using "cv.hqreg" funcction from "hqreg" package and "rqPen" means using "rq.pen" function with "aLASSO" penalty.}

\item{seed}{seed for hqreg estimation process.}
}
\value{
A list.
\itemize{
    \item ts - Test statistic.
    \item pval - p value.
    \item tr_sigma - trace sigma square.
    \item coef - the coefficient of control factors.
}
}
\description{
High dimensional quantile partial test
}
\examples{
set.seed(0)
n <- 300
p <- 710
x <- matrix(rnorm(n * p), n, p)
u <- matrix(rnorm(n * p), n, p)
alpha <- c(rep(1, 5), rep(0, p - 5))
beta <- c(rep(1, 5), rep(0, p - 5))
y <-  u \%*\% alpha + x \%*\% beta + rnorm(n)
test_result <- hdqpt(x, u, y, tau = 0.5, decorrelate = FALSE, method = "hqreg")
print(test_result)
}
